NXP-29268: make upload.sh not crash if there is no package
Now that the last package Web UI was removed, no package is being
upload.

+++ upload.sh
@@ -20,7 +20,9 @@ if [ ! -z "$UPLOAD_URL" ]; then
     for pkg in $(cat .featured); do
         scp -C $pkg*.zip $UPLOAD_URL/mp-nuxeo-server/ || true
     done
-    scp packages.xml $UPLOAD_URL/mp-nuxeo-server/ || true
+    if [ -f "packages.xml" ]; then
+        scp packages.xml $UPLOAD_URL/mp-nuxeo-server/ || true
+    fi
     cd -
     date
 fi
